Revert "update eorder project"
This reverts commit d839522f73f5f5ac4b63d5084f1c1d77a03c1392.

+++ 02_code/web/eorder/src/main/java/com/innovaee/eorder/entity/UserLevel.java
@@ -32,7 +32,7 @@ public class UserLevel extends BaseEntity {
     private Float discount;
 
     /** 等级积分 */
-    private Integer levelScore;
+    private Long levelScore;
 
     /** 等级状态 */
     private Boolean levelStatus;
@@ -59,11 +59,11 @@ public class UserLevel extends BaseEntity {
 
     @Basic
     @Column(name = "level_score")
-    public Integer getLevelScore() {
+    public Long getLevelScore() {
         return levelScore;
     }
 
-    public void setLevelScore(Integer levelScore) {
+    public void setLevelScore(Long levelScore) {
         this.levelScore = levelScore;
     }
